rename logentry fixture module
