+++ tests/fixtures/core/__init__.py
@@ -5,11 +5,11 @@ from tests.fixtures.core.api import as_user
 from tests.fixtures.core.current_user import mock_order_refund_service_current_user
 from tests.fixtures.core.factory import factory
 from tests.fixtures.core.factory import mixer
-from tests.fixtures.core.logentry import write_admin_log
 from tests.fixtures.core.pause_slow_background_jobs import _pause_dashamail
 from tests.fixtures.core.pause_slow_background_jobs import _pause_tags
 from tests.fixtures.core.send_mail import send_mail
 from tests.fixtures.core.timezone import kamchatka_timezone
+from tests.fixtures.core.write_admin_log import write_admin_log
 
 __all__ = [
     "anon",
