+++ tests/fixtures/core/write_admin_log.py
